Moves all but DIstricts to standalone classes. Need to add districts and watepipe plotter, HeatPipe converter, not sure if allow tunnels warrents moving or not, same with precision engineering
