+++ ARUT.cs
@@ -26,93 +26,6 @@ namespace AnotherRoadUpdateTool
     {
         #region Declarations
 
-        #region enums
-
-        private enum tp
-        {
-            Ground = 0,
-            Bridge = 1,
-            Slope = 2,
-            Tunnel = 3,
-            Curve = 4
-        }
-
-        private enum Up
-        {
-            Basic = 0,
-            Highway = 1,
-            Large = 2,
-            Medium = 3,
-            Oneway = 4
-        }
-
-        private enum p
-        {
-            Roads = 0,
-            Railroads = 1,
-            Highways = 2,
-            PowerLines = 3,
-            WaterPipes = 4,
-            HeatPipes = 5,
-            Airplanes = 6,
-            Shipping = 7,
-            Pedestrian = 8,
-            Bicycle = 9,
-            Tram = 10,
-            Metro = 11,
-            Buildings = 12,
-            Trees = 13,
-            Props = 14
-        }
-
-        private enum ops
-        {
-            Updates = 0,
-            Deletes = 1,
-            Services = 2,
-            Terrain = 3,
-            Districts = 4
-        }
-
-        private enum dl
-        {
-            HealthCare = 0,
-            PoliceDepartment = 1,
-            FireDepartment = 2,
-            PublicTransport = 3,
-            Education = 4,
-            Electricity = 5,
-            Water = 6,
-            Garbage = 7,
-            Beautification = 8,
-            Monument = 9,
-            Abandoned = 10,
-            Burned = 11,
-            Chirper = 12
-        }
-
-        //None = 0,
-        //Residential = 1,
-        //Commercial = 2,
-        //Industrial = 3,
-        //Unused1 = 4,
-        //Unused2 = 5,
-        //Citizen = 6,
-        //Tourism = 7,
-        //Office = 8,
-        //Road = 9,
-        //Electricity = 10,
-        //Water = 11,
-        //Beautification = 12,
-        //Garbage = 13,
-        //HealthCare = 14,
-        //PoliceDepartment = 15,
-        //Education = 16,
-        //Monument = 17,
-        //FireDepartment = 18,
-        //PublicTransport = 19,
-        //Government = 20
-
         public struct UndoStroke
         {
             public string name;
@@ -125,8 +38,6 @@ namespace AnotherRoadUpdateTool
             public ushort[] rawHeights;
             public int pointer;
         }
-
-        #endregion
 
         #region Variables
 
@@ -155,10 +66,16 @@ namespace AnotherRoadUpdateTool
         internal static bool AllRoads;
         internal static bool AutoDistroy;
 
+        internal static bool ShowDelete;
+        internal static bool ShowUpdate;
+        internal static bool ShowTerrain;
+        internal static bool ShowServices;
+        internal static bool ShowDistricts;
+
         internal static int MaxAreas;
         //internal static int StartMoney;
         internal static int StartMoney { get { return us.StartMoney; } private set { value = us.StartMoney; }  }
-
+        
         internal readonly static SavedBool DemolishAbandoned = new SavedBool("ModDemolishAbandoned", Settings.gameSettingsFile, true, true);
         internal readonly static SavedBool DemolishBurned = new SavedBool("ModDemolishBurned", Settings.gameSettingsFile, true, true);
 
@@ -171,7 +88,7 @@ namespace AnotherRoadUpdateTool
         private object m_dataLock = new object();
 
         private LoadMode mode;
-        private Chirper Chirp;
+        private static Chirper Chirp;
         private UnlockRoads Roads;
         private DestroyMonitor Dozer;
         private MaxAreas Areas;
@@ -183,29 +100,8 @@ namespace AnotherRoadUpdateTool
 
         #region not Arrays
 
-        private bool ShowDelete;
-        private bool ShowUpdate;
-        private bool ShowTerrain;
-        private bool ShowServices;
-        private bool ShowDistricts;
-
         private bool m_active;
-        private bool m_selectable;
-        private bool m_fromTypes;
-        private bool m_toTypes;
-        private bool m_toRoads;
-        private bool m_fromRoads;
-        private bool m_updates;
         private bool m_settings;
-        private bool mouseDown;
-
-        private string m_available = "available.";
-        private string m_unavailable = "unavailable.";
-        private string m_updatetool = "Another Road Update Tool - Selection is ";
-        private string m_defaultInfo = "Watch for additional information.";
-
-        private string fromSelected = string.Empty;
-        private string toSelected = string.Empty;
 
         private Position PanelPosition;
         private Vector3 m_startPosition;
@@ -218,10 +114,6 @@ namespace AnotherRoadUpdateTool
         private new Ray m_mouseRay;
         private new float m_mouseRayLength;
 
-        private float m_terrainHeight = 0.0f;
-
-        private
-
         readonly ushort[] m_undoBuffer = Singleton<TerrainManager>.instance.UndoBuffer;
         private ushort[] m_originalHeights;
         readonly ushort[] m_backupHeights = Singleton<TerrainManager>.instance.BackupHeights;
@@ -231,103 +123,10 @@ namespace AnotherRoadUpdateTool
         
         private float m_maxArea = 400f;
 
-        private UIPanel plMain;
-        private UIPanel plOptions;
-        private UIPanel plTypes;
-        private UIPanel plRoads;
-        private UIPanel plDelete;
-        private UIPanel plServices;
-        private UIPanel plBasic;
-        private UIPanel plHighway;
-        private UIPanel plLarge;
-        private UIPanel plMedium;
-        private UIPanel plOneway;
-        private UIPanel plToBasic;
-        private UIPanel plToHighway;
-        private UIPanel plToLarge;
-        private UIPanel plToMedium;
-        private UIPanel plToOneway;
-        private UIPanel plTerrain;
-        private UIPanel plDistricts;
-
-        private UILabel lLines;
-        private UILabel lProperties;
-        private UILabel lSelectable;
-        private UILabel lInformation;
-        private UILabel lHeight;
-        private UILabel lbTitle;
-
-        private UIButton btHelp;
-        private UIButton btHide;
-        private UIButton btValidate;
-        private UIButton mainButton;
-        private UICheckBox cbToggle;
-        private UICheckBox cbDistrictToggle;
-        private UITextField tfTerrainHeight;
-
-        private List<UICheckBox> panels = new List<UICheckBox>();
-        private List<UICheckBox> options = new List<UICheckBox>();
-        private List<UICheckBox> deletes = new List<UICheckBox>();
-        private List<UICheckBox> services = new List<UICheckBox>();
-        private List<UICheckBox> types = new List<UICheckBox>();
-
-        private List<UICheckBox> toTypes = new List<UICheckBox>();
-        private List<UICheckBox> fromTypes = new List<UICheckBox>();
-
-        private List<UICheckBox> toBasic = new List<UICheckBox>();
-        private List<UICheckBox> fromBasic = new List<UICheckBox>();
-
-        private List<UICheckBox> toLarge = new List<UICheckBox>();
-        private List<UICheckBox> fromLarge = new List<UICheckBox>();
-
-        private List<UICheckBox> toHighway = new List<UICheckBox>();
-        private List<UICheckBox> fromHighway = new List<UICheckBox>();
-
-        private List<UICheckBox> toMedium = new List<UICheckBox>();
-        private List<UICheckBox> fromMedium = new List<UICheckBox>();
-
-        private List<UICheckBox> toOneway = new List<UICheckBox>();
-        private List<UICheckBox> fromOneway = new List<UICheckBox>();
+        private PanelMain plMain;
+        internal UIButton mainButton;
 
         #endregion
-
-        #region String Arrays
-
-        //These strings are importent in that they control the interface
-        private string[] m_options = new string[] { "Updates", "Deletes", "Services", "Terrain", "Districts" };
-
-        private string[] m_types = new string[] { "Ground", "Bridge", "Slope", "Tunnel", "Curve" };
-        private string[] m_roads = new string[] { "Label ToFrom", "Basic", "Highway", "Large", "Medium", "Oneway" };
-        private string[] m_basic = new string[] { "Basic Road", "Basic Road Decoration Grass", "Basic Road Decoration Trees",
-            "Basic Road Bicycle", "Basic Road Tram", "Gravel Road" };
-
-        private string[] m_highway = new string[] { "Highway Barrier", "Highway", "HighwayRamp" };
-
-        private string[] m_large = new string[] { "Large Road", "Large Road Decoration Grass", "Large Road Decoration Trees",
-            "Large Road Bicycle", "Large Road Bus", "Large Oneway", "Large Oneway Decoration Grass", "Large Oneway Decoration Trees" };
-
-        private string[] m_medium = new string[] { "Medium Road", "Medium Road Decoration Grass", "Medium Road Decoration Trees",
-            "Medium Road Bicycle", "Medium Road Bus", "Medium Road Tram" };
-
-        private string[] m_oneway = new string[] { "Oneway Road", "Oneway Road Decoration Grass", "Oneway Road Decoration Trees",
-            "Oneway Road Tram", "Tram Track", "Oneway Tram Track" };
-
-        private string[] m_deletes = new string[] { "Label Lines", "Roads", "B Railroads", "Highways", "B PowerLines", "Water Pipes",
-            "B Heat Pipes", "Airplanes", "B Shipping", "Pedestrian", "B Bicycle", "Tram", "B Metro", "Label Properties", "Buildings",
-            "Trees", "Props" };
-
-        private string[] m_services = new string[] {"Label Select Toggle Services", "HealthCare", "PoliceDepartment", "FireDepartment",
-            "PublicTransport", "Education", "Electricity", "Water", "Garbage", "Beautification", "Monument", "Select Options",
-            "Auto Bulldozer",  "Chirper" };
-
-        private string[] m_heights = new string[] { "0.00", "5.00", "10.00", "15.00", "20.00", "25.00", "30.00", "35.00", "40.00",
-            "45.00", "50.00", "55.00", "60.00", "65.00", "70.00", "75.00", "80.00", "85.00", "90.00", "95.00", "100.00", "150.00",
-            "200.00", "250.00", "300.00", "350.00", "400.00", "450.00", "500.00", "550.00", "600.00", "650.00", "700.00", "750.00",
-            "800.00", "850.00", "900.00", "950.00", "1000.00", "1500.00", "2000.00" };
-
-        private string[] m_incraments = new string[] { "1", "10", "50", "100", "250", " 500" };
-
-        #endregion //String Arrays
 
         #endregion
 
@@ -357,21 +156,6 @@ namespace AnotherRoadUpdateTool
             base.OnDestroy();
         }
 
-        protected override void OnEnable()
-        { 
-            //WriteLog("ARUT OnEnable!");
-            base.OnEnable();
-        }
-
-        protected override void OnDisable()
-        {
-            if (plMain != null)
-                plMain.isVisible = false;
-
-            //WriteLog("ARUT OnDisable!");
-            base.OnDisable();
-        }
-
         protected override void OnToolUpdate()
         {
             //WriteLog("ARUT OnToolUpdate!");
@@ -398,7 +182,7 @@ namespace AnotherRoadUpdateTool
                 }
             }
 
-            if (e.type == EventType.MouseDown && m_mouseRayValid && m_selectable)
+            if (e.type == EventType.MouseDown && m_mouseRayValid && plMain.Selectable)
             {
                 if (e.button == 0)
                 {
@@ -411,46 +195,48 @@ namespace AnotherRoadUpdateTool
                     m_active = false;
                 }
             }
-            else if (e.type == EventType.MouseUp && m_active)
+            else if (e.type == EventType.MouseUp && m_active && plMain.Selectable)
             {
                 if (e.button == 0)
                 {
                     //handle Updates
-                    if (options[(int)ops.Updates].isChecked && plTypes.isVisible)
+                    if (plMain.Options(PanelMain.ops.Updates) && plMain.plTypes.isVisible)
                     {
-                        UpdateObjects uo = new UpdateObjects(lInformation, m_startPosition, m_mousePosition, deletes, types, fromSelected, toSelected);
+                        UpdateObjects uo = new UpdateObjects(plMain.lInformation, m_startPosition, m_mousePosition, plMain.deletes, plMain.types, plMain.FromSelected, plMain.ToSelected);
                         Thread t = new Thread(new ThreadStart(uo.ApplyUpdates));
                         t.Start();
                         while (!t.IsAlive);
                     }
                     //handle Deletes
-                    else if (options[(int)ops.Deletes].isChecked && plDelete.isVisible)
+                    else if (plMain.Options(PanelMain.ops.Deletes) && plMain.plDelete.isVisible)
                     {
-                        DeleteObjects rd = new DeleteObjects(m_startPosition, m_mousePosition, deletes, types);
+                        DeleteObjects rd = new DeleteObjects(m_startPosition, m_mousePosition, plMain.deletes, plMain.types);
                         Thread t = new Thread(new ThreadStart(rd.ApplyDeletes));
                         t.Start();
                         while (!t.IsAlive);
                     }
                     //handle Services
-                    else if (options[(int)ops.Services].isChecked && plServices.isVisible)
+                    else if (plMain.Options(PanelMain.ops.Services) && plMain.plServices.isVisible)
                     {
-                        if (mode != LoadMode.LoadMap && mode != LoadMode.NewMap)
-                            ApplyServices();
+                        ToggleServices ts = new ToggleServices(plMain, m_startPosition, m_mousePosition);
+                        Thread t = new Thread(new ThreadStart(ts.ApplyServices));
+                        t.Start();
+                        while (!t.IsAlive);
                     }
                     //handle Districts
-                    else if (options[(int)ops.Districts].isChecked && plDistricts.isVisible)
+                    else if (plMain.Options(PanelMain.ops.Districts) && plMain.plDistricts.isVisible)
                     {
                         WriteLog("Trying ApplyDistrictsChange");
                         ApplyDistrictsChange();
                         WriteLog("Tried ApplyDistrictsChange");
                     }
                     //handle Terrain
-                    else if (options[(int)ops.Terrain].isChecked && plTerrain.isVisible)
+                    else if (plMain.Options(PanelMain.ops.Terrain) && plMain.plTerrain.isVisible)
                     {
                         this.m_endPosition = this.m_mousePosition;
                         if (mode == LoadMode.LoadMap || mode == LoadMode.NewMap)
                         {
-                            TerrainChanges tc = new TerrainChanges(m_startPosition, m_endPosition, m_terrainHeight, m_originalHeights, m_backupHeights, m_rawHeights);
+                            TerrainChanges tc = new TerrainChanges(m_startPosition, m_endPosition, (float)plMain.TerrainHeight, m_originalHeights, m_backupHeights, m_rawHeights);
                             Thread t = new Thread(new ThreadStart(tc.ApplyTerrainChange));
                             t.Start();
                             while (!t.IsAlive) ;
@@ -516,686 +302,47 @@ namespace AnotherRoadUpdateTool
 
                 mainButton.eventClick += Button_Clicked;
 
-                plMain = UIView.GetAView().FindUIComponent("TSBar").AddUIComponent<UIPanel>();
+                WriteLog("About to set PanelMain in InitGUI");
+                GameObject go = new GameObject("buildingWindowObject");
 
-                plMain.anchor = UIAnchorStyle.None;
+                plMain = go.AddComponent<PanelMain>();
 
-                if (mode == LoadMode.LoadMap || mode == LoadMode.NewMap)
-                    plMain.backgroundSprite = "GenericPanel";
-                else
-                    plMain.backgroundSprite = "SubcategoriesPanel";
+                var view = UIView.GetAView();
 
-                int height = 525;
+                plMain.Mode = mode;
+                plMain.transform.parent = view.transform;
+                plMain.isVisible = true;
+                plMain.canFocus = true;
+                plMain.isInteractive = true;
+                plMain.relativePosition = new Vector3(572, 525);
 
-                plMain.size = new Vector2(575, height);
+                plMain.backgroundSprite = "MenuPanel2";
+                plMain.CreateObjects();
 
-                lbTitle = addLabel(plMain, 1, 1, "ARUT - Another 'Road' Update Tool", "All your Mods in one control", false);
-                lbTitle.autoSize = false;
-                lbTitle.isVisible = true;
-                lbTitle.outlineColor = Color.red;
-                lbTitle.outlineSize = 2;
-                lbTitle.bottomColor = Color.yellow;
-                lbTitle.width = plMain.width;
-                lbTitle.height =  20;
-                lbTitle.textScale = 0.9f;
-                lbTitle.textAlignment = UIHorizontalAlignment.Center;
-                lbTitle.verticalAlignment = UIVerticalAlignment.Middle;
-
-                string tooltip = "I will try to open a pdf file, if you do not have a viewer.... do not click.";
-                btHelp = addButton(plMain, "¿", tooltip, (int)plMain.height - 25, (int)plMain.width - 25, 25, 25);
-                btHelp.isVisible = true;
-                //btHelp.zOrder = 0;
-                btHelp.eventDoubleClick += btHelp_eventDoubleClick;
-
-                tooltip = "I will try to close the option panel.";
-                btHide = addButton(plMain, "Close", tooltip, (int)plMain.height - 25, 1, 75, 25);
-                btHide.isVisible = true;
-                //btHide.zOrder = 0;
-                btHide.eventClick += btHide_eventClick;
-
-                plMain.eventMouseDown += Main_eventMouseDown;
-                plMain.eventMouseMove += Main_eventMouseMove;
-                plMain.eventMouseUp += Main_eventMouseUp;
-                plMain.eventSizeChanged += Main_eventSizeChanged;
-
-                plMain.relativePosition = new Vector2
-                (
-                    RoadUpdateButton.relativePosition.x + RoadUpdateButton.width / 2.0f - plMain.width,
-                    RoadUpdateButton.relativePosition.y - (plMain.height + 3)
-                );
-
-                //Create the panels (Little like a tab view)
-                height = CreatePanels(plMain);
-
+                WriteLog("About to set GetSettings in InitGUI");
                 //We can load the users last session
                 GetSettings();
 
+                plMain.RefreshView();
+
+                WriteLog("About to set Areas and Zones in InitGUI");
+
                 //About to set unlockable tiles
                 Areas = new MaxAreas();
-                Zones = new Helpers.Zones();
+                Zones = new Zones();
                 //WriteLog("Leaving InitGUI");
             }
+        }
+
+        internal static void SetChirper(bool value)
+        {
+            if (Chirp.ToggleState != value)
+                Chirp.Toggle(value);
         }
 
         #endregion
 
         #region "Private procedures"
-
-        #region "GUI Layout"
-
-        #region Panels
-
-        private int CreatePanels(UIPanel panel)
-        {
-            int plx = 1;
-            int options = 20;
-            int services = 0;
-            int terrain = 0;
-            int types = 0;
-            int roads = 0;
-
-            services = GenerateOptions(panel, options, plx);    //Options
-            types = GenerateplTypes(panel, services, plx);      //Types
-            roads = GenerateplRoads(panel, types, plx);         //Updates
-            GenerateplDelete(panel, types, plx);                //Deletes
-            GenerateplServices(panel, services, plx);           //Services
-            GenerateplTerrain(panel, services, plx);            //Terrain
-            GenerateplDistricts(panel, terrain, plx);           //Districts
-
-            GenerateRoadPanels(panel, ref plBasic, ref plToBasic, fromBasic, toBasic, m_basic, "Basic", roads, plx);
-            GenerateRoadPanels(panel, ref plHighway, ref plToHighway, fromHighway, toHighway, m_highway, "Highway", roads, plx);
-            GenerateRoadPanels(panel, ref plLarge, ref plToLarge, fromLarge, toLarge, m_large, "Large", roads, plx);
-            GenerateRoadPanels(panel, ref plMedium, ref plToMedium, fromMedium, toMedium, m_medium, "Medium", roads, plx);
-            GenerateRoadPanels(panel, ref plOneway, ref plToOneway, fromOneway, toOneway, m_oneway, "Oneway", roads, plx);
-
-            WriteLog("services = " + services);
-            WriteLog("types = " + types);
-            WriteLog("terrain = " + terrain);
-            WriteLog("deletes = " + deletes);
-            WriteLog("roads = " + roads);
-
-            WriteLog("plOptions.size = " + plOptions.size);
-            WriteLog("plTypes.size = " + plTypes.size);
-            WriteLog("plRoads.size = " + plRoads.size);
-            WriteLog("plServices.size = " + plServices.size);
-            WriteLog("plTerrain.size = " + plTerrain.size);
-            WriteLog("plDistricts.size = " + plDistricts.size);
-            WriteLog("plMain.size = " + panel.size);
-
-            WriteLog("plOptions.position = " + plOptions.position);
-            WriteLog("plTypes.position = " + plTypes.position);
-            WriteLog("plRoads.position = " + plRoads.position);
-            WriteLog("plServices.position = " + plServices.position);
-            WriteLog("plTerrain.position = " + plTerrain.position);
-            WriteLog("plDistricts.position = " + plDistricts.position);
-            WriteLog("plMain.position = " + panel.position);
-            
-            return 525;
-        }
-
-        private int GenerateOptions(UIPanel panel, int ply, int plx)
-        {
-            plOptions = panel.AddUIComponent<UIPanel>();
-            plOptions.relativePosition = new Vector3(plx, ply);
-            plOptions.isVisible = true;
-            plOptions.tooltip = "Select the type of updates and options to perform.";
-
-            int y = 1;
-            lSelectable = addLabel(plOptions, y, plx, m_updatetool + m_unavailable, true);
-            y += 25;
-            lInformation = addLabel(plOptions, y, 1, m_defaultInfo, true);
-            y += 25;
-
-            int cb = 0;
-            foreach (string s in m_options)
-            {
-                bool enable = true;
-                string t = String.Format("Select to display the {0} options", s);
-                options.Add(addCheckbox(plOptions, y, plx, s, t, true));
-                //Space out the options (We may ad building, trees, and props)
-                plx += 100;
-                switch (s)
-                {
-                    case "Update":
-                        enable = us.ShowUpdate;
-                        break;
-                    case "Delete":
-                        enable = us.ShowDelete;
-                        break;
-                    case "Districts":
-                        enable = (us.ShowDistricts == (mode != LoadMode.LoadMap && mode != LoadMode.NewMap));
-                        break;
-                    case "Terrain":
-                        enable = (us.ShowTerrain == (mode == LoadMode.LoadMap || mode == LoadMode.NewMap));
-                        break;
-                    case "Services":
-                        enable = (us.ShowServices == (mode != LoadMode.LoadMap && mode != LoadMode.NewMap));
-                        break;
-                    default:
-                        break;
-                }
-                options[cb].enabled = enable;
-                options[cb].eventCheckChanged += Options_eventCheckChanged;
-                cb += 1;
-            }
-            y += 25;
-
-            //set the panal size (two rows, 50)
-            plOptions.size = new Vector2(panel.width, y);
-            
-            return (int)plOptions.height + 20;
-        }
-
-        private int GenerateplTypes(UIPanel panel, int ply, int plx)
-        {
-            plTypes = panel.AddUIComponent<UIPanel>();
-            plTypes.relativePosition = new Vector3(1, ply);
-            plTypes.isVisible = true;
-            plTypes.tooltip = "Select the line types to modify";
-
-            int y = 1;
-            addLabel(plTypes, y, 1, "Select the line types to modify", true);
-            y += 25;
-            int x = 5;
-            int step = 0;
-            int cb = 0;
-
-            //load the update types
-            foreach (string s in m_types)
-            {
-                string t = "These types will be updated.";
-                types.Add(addCheckbox(plTypes, y, x + step, s, t, true));
-                types[cb].eventCheckChanged += Types_eventCheckChanged;
-                step += 100;
-                cb += 1;
-            }
-            //set the panel size (two rows, 50)
-            y += 20;
-            plTypes.size = new Vector2(panel.width, y);
-
-            return ply + y;
-        }
-
-        private int GenerateplRoads(UIPanel panel, int ply, int plx)
-        {
-            //WriteLog("Entering GenerateplTypes");
-            //Show the road type option
-            plRoads = panel.AddUIComponent<UIPanel>();
-            plRoads.relativePosition = new Vector3(1, ply);
-            plRoads.isVisible = false;
-            plRoads.tooltip = "Select the Road types to convert from and to";
-
-            int x1 = 10;
-            int x2 = 280;
-            int y = 1;
-            int cb = 0;
-            int step = 0;
-            //load the update types but hide them
-            foreach (string s in m_roads)
-            {
-                if (s.StartsWith("Label "))
-                {
-                    try
-                    {
-                        addLabel(plRoads, y, x1 - 5, "From ==> ", true);
-                        addLabel(plRoads, y, x2 - 5, "To ==> ", true);
-                        y += 20;
-                    }
-                    catch (Exception ex)
-                    {
-                        WriteError("Error in GenerateplTypes: ", ex);
-                    }
-                }
-                else
-                {
-                    string t = "";
-                    if (s == "Ground" || s == "Bridge" || s == "Tunnel" || s == "Slope" || s == "Curve")
-                    {
-                        t = "These types will be updated.";
-                        deletes.Add(addCheckbox(plDelete, y, x1 + step, s, t, true));
-                        step += 110;
-                        if (s == "Curve")
-                            y += 25;
-                    }
-                    else
-                    {
-                        t = String.Format("If checked {0} sections will be displayed.", s);
-                        fromTypes.Add(addCheckbox(plRoads, y, x1, s, t, true));
-                        toTypes.Add(addCheckbox(plRoads, y, x2, s, t, true));
-                        fromTypes[cb].eventCheckChanged += FromTypes_eventCheckChanged;
-                        toTypes[cb].eventCheckChanged += ToTypes_eventCheckChanged;
-                    }
-                    y += 25;
-                    cb++;
-                }
-            }
-            plRoads.size = new Vector2(panel.width, 200);
-            //WriteLog("Leaving GenerateplTypes");
-            //return the top of the roads panals
-            return ply + y;
-        }
-
-        private void GenerateplDelete(UIPanel panel, int ply, int plx)
-        {
-            //WriteLog("Entering GenerateplDelete");
-            //Show the road type option
-            plDelete = panel.AddUIComponent<UIPanel>();
-            plDelete.relativePosition = new Vector3(1, ply);
-            plDelete.isVisible = false;
-            plDelete.tooltip = "Select the type of items to delete.";
-
-            int cb = 0;
-            int x = 15;
-            int y = 1;
-
-            //load the bulldoze road type options
-            addLabel(plDelete, y, 5, "Select your delete options.", true);
-
-            y += 20;
-            foreach (string s in m_deletes)
-            {
-                if (s == "Label Lines")
-                {
-                    lLines = addLabel(plDelete, y, 10, "Lines", true);
-                    y += 20;
-                }
-                else if (s == "Label Properties")
-                {
-                    lProperties = addLabel(plDelete, y, 10, "Properties", true);
-                    y += 20;
-                }
-                else
-                {
-                    string t = "These items will be deleted.";
-                    if (s.StartsWith("B "))
-                    {
-                        deletes.Add(addCheckbox(plDelete, y, x + 210, s.Replace("B ", ""), t, true));
-                        y += 25;
-                    }
-                    else
-                    {
-                        deletes.Add(addCheckbox(plDelete, y, x, s, t, true));
-                        if (s == "Buildings" || s == "Trees" || s == "Props") { y += 25; }
-                    }
-                    deletes[cb].eventCheckChanged += DeleteTypes_eventCheckChanged;
-                    cb += 1;
-                }
-            }
-
-            plDelete.size = new Vector2(panel.width, y + 25);
-            //WriteLog("Leaving GenerateplDelete: " + deletes.Count);
-        }
-
-        private void GenerateplServices(UIPanel panel, int ply, int plx)
-        {
-            //WriteLog("Entering GenerateplServices");
-            //Show the road type option
-            plServices = panel.AddUIComponent<UIPanel>();
-            plServices.relativePosition = new Vector3(1, ply);
-            plServices.isVisible = false;
-            plServices.tooltip = "Select the type of services to toggle on and off.";
-
-            int cb = 0;
-            int y = 1;
-
-            for ( int i = 0; i < m_services.Length; i++)
-            {
-                string s = m_services[i];
-
-                string t = "These items will be toggled On or Off.";
-                
-                if (s == "Auto Bulldozer")
-                {
-                    addLabel(plServices, y, 10, s, true);
-                    y += 25;
-                    services.Add(addCheckbox(plServices, y, 15, "Abandoned", "Check to delete abandoned buildings.", true));
-                    services[cb].eventCheckChanged += ServiceTypes_eventCheckChanged;
-                    cb += 1;
-                    services.Add(addCheckbox(plServices, y, 150, "Burned", "Check to delete burned buildings.", true));
-                    services[cb].eventCheckChanged += ServiceTypes_eventCheckChanged;
-                    cb += 1;
-                }
-                else if (s == "Select Options")
-                {
-                    addLabel(plServices, y, 5, s, true);
-                }
-                else if (s == "Chirper")
-                {
-                    addLabel(plServices, y, 10, s, true);
-                    y += 25;
-                    services.Add(addCheckbox(plServices, y, 15, s, "Checked to Show, unchecked to Hide Chirper.", true));
-                    services[cb].eventCheckChanged += ServiceTypes_eventCheckChanged;
-                    cb += 1;
-                }
-                else if (s.StartsWith("Label "))
-                {
-                    addLabel(plServices, y, 5, s.Replace("Label ", ""), true);
-                    t = "Check to turn on services, unchecked to turn them off.";
-                    y += 25;
-                    cbToggle = addCheckbox(plServices, y, 10, "Check to turn Sevices On, Uncheck for Off.", t, true);
-                }
-                else
-                {
-                    s = m_services[i];
-                    services.Add(addCheckbox(plServices, y, 15, s, t, true));
-                    services[cb].eventCheckChanged += ServiceTypes_eventCheckChanged;
-                    i++;
-                    cb += 1;
-                    if (i < m_services.Length)
-                    {
-                        s = m_services[i];
-                        services.Add(addCheckbox(plServices, y, 285, s, t, true));
-                        services[cb].eventCheckChanged += ServiceTypes_eventCheckChanged;
-                        cb += 1;
-                    }
-                }
-                //WriteLog("Value and Index: " + s + ", " + cb);
-                y += 25;
-            }
-            //no need for this dropdown
-            //ServiceDropdown(plServices);
-            plServices.size = new Vector2(panel.width, y + 25);
-            //WriteLog("Leaving GenerateplServices");
-        }
-
-        private void GenerateplTerrain(UIPanel panel, int ply, int plx)
-        {
-            //WriteLog("Entering GenerateplTerrain");
-            
-            //Show the road type option
-            plTerrain = panel.AddUIComponent<UIPanel>();
-            plTerrain.relativePosition = new Vector3(1, ply);
-            plTerrain.isVisible = false;
-            plTerrain.tooltip = "Select or enter the height desired.";
-
-            int x = 1;
-            int y = 1;
-            int inset = 65;
-
-            string s = "Enter a value for the new terrain height";
-            string t = "Enter the height desired.";
-            addLabel(plTerrain, y, x, s, t, true);
-
-            y += 25;
-            s = "Current: ";
-            t = "Look here for current new Terrain height.";
-            addLabel(plTerrain, y, x, s, t, true);
-
-            s = "0.00";
-            lHeight = addLabel(plTerrain, y, x + inset, s, t, true);
-            lHeight.autoSize = false;
-            lHeight.width = 100;
-            lHeight.height = 50;
-            lHeight.textAlignment = UIHorizontalAlignment.Right;
-
-            y += 25;
-            s = "Height";
-            addLabel(plTerrain, y, x, s, t, true);
-
-            tfTerrainHeight = addTextBox(plTerrain, "TerrainHeight", "0,00", y, x + inset, 100, 25, "Use values between 2000 and 0.0", true, true);
-            tfTerrainHeight.horizontalAlignment = UIHorizontalAlignment.Right;
-            tfTerrainHeight.eventKeyDown += TerrainHeight_eventKeyDown;
-            tfTerrainHeight.eventTextChanged += TerrainHeight_eventTextChanged;
-
-            btValidate = addButton(plTerrain, "Validate", "Validate the number in the text field.", y, x + inset + 105, 125, 25);
-            btValidate.isVisible = true;
-            btValidate.state = UIButton.ButtonState.Disabled;
-            btValidate.eventClick += Validate_eventClick;
-
-            y += 25;
-            s = "Game Height";
-            t = "Click to load in the current Games set Terrain Height";
-            UIButton GameHeight = addButton(plTerrain, s, t, y, x, 100, 25);
-            GameHeight.eventClick += GameHeight_eventClick;
-        }
-        
-        private void GenerateplDistricts(UIPanel panel, int ply, int plx)
-        {
-            //WriteLog("Entering GenerateplDistricts");
-
-            //Show the road type option
-            plDistricts = panel.AddUIComponent<UIPanel>();
-            plDistricts.relativePosition = new Vector3(1, ply);
-            plDistricts.isVisible = false;
-            plDistricts.tooltip = "Select an area to update the district.";
-
-            addLabel(plDistricts, 1, 1, "Toggle the check box to create or delete the district. (Undo is not available)", true);
-            addLabel(plDistricts, 20, 1, "Add to an existing District by starting the selection within the district.", true);
-
-            cbDistrictToggle = addCheckbox(plDistricts, 40, 15, "Checked for Create (Update), unchecked for delete", "Check to create a district, uncheck to delete", true);
-
-           // WriteLog("Leaving GenerateplDistricts");
-        }
-
-        private void GenerateRoadPanels(UIPanel panel, ref UIPanel pFrom, ref UIPanel pTo, List<UICheckBox> lFrom, List<UICheckBox> lTo, string[] road, string name, int ply, int plx)
-        {
-            //WriteLog("Entering GenerateRoadPanels");
-            //Show the road from option
-            pFrom = panel.AddUIComponent<UIPanel>();
-            pFrom.relativePosition = new Vector3(1, ply);
-            pFrom.size = new Vector2(25 * 12, panel.width / 2);
-            pFrom.isVisible = false;
-            pFrom.tooltip = "Select the type of road to convert.";
-            //Show the road to option
-            pTo = panel.AddUIComponent<UIPanel>();
-            pTo.relativePosition = new Vector3(285, ply);
-            pTo.size = new Vector2(25 * 12, panel.width / 2);
-            pTo.isVisible = false;
-            pTo.tooltip = "Select the type of road to convert to.";
-
-            addLabel(pFrom, 1, 5, pFrom.tooltip, true);
-            addLabel(pTo, 1, 5, pTo.tooltip, true);
-
-            int cb = 0;
-            int y = 22;
-            //load the update roads but hide them
-            foreach (string s in road)
-            {
-                string t = String.Format("If checked {0} sections will be converted.", s);
-                lFrom.Add(addCheckbox(pFrom, y, 11, s, t, true));
-                t = String.Format("If checked sections will convert to {0}.", s);
-                lTo.Add(addCheckbox(pTo, y, 11, s, t, true));
-                lFrom[cb].eventCheckChanged += FromRoad_eventCheckChanged;
-                lTo[cb].eventCheckChanged += ToRoad_eventCheckChanged;
-                cb++;
-                y += 25;
-            }
-            //WriteLog("Leaving GenerateRoadPanels");
-        }
-
-        #endregion
-
-        #region "Adding Controls"
-
-        private UILabel addLabel(UIPanel panel, int yPos, int xPos, string text, bool hidden)
-        {
-            return addLabel(panel, yPos, xPos, text, "", hidden);
-        }
-
-        private UILabel addLabel(UIPanel panel, int yPos, int xPos, string text, string t, bool hidden)
-        {
-            //WriteLog("Entering addUILabel");
-            UILabel lb = panel.AddUIComponent<UILabel>();
-            lb.relativePosition = new Vector3(xPos, yPos);
-            lb.height = 0;
-            lb.width = 80;
-            lb.text = text;
-            lb.tooltip = t;
-            lb.isVisible = hidden;
-            //WriteLog("Leaving addUILabel");
-            return lb;
-        }
-
-        private UICheckBox addCheckbox(UIPanel panel, int yPos, int xPos, string text, string tooltip, bool hidden)
-        {
-            //WriteLog("Entering addCheckbox");
-            var cb = panel.AddUIComponent<UICheckBox>();
-            cb.relativePosition = new Vector3(xPos, yPos);
-            cb.height = 0;
-            cb.width = 80;
-
-            var label = panel.AddUIComponent<UILabel>();
-            label.relativePosition = new Vector3(xPos + 25, yPos + 3);
-            cb.label = label;
-            cb.label.tabIndex = cb.GetInstanceID();
-            cb.text = text;
-
-            UISprite uncheckSprite = cb.AddUIComponent<UISprite>();
-            uncheckSprite.height = 20;
-            uncheckSprite.width = 20;
-            uncheckSprite.relativePosition = new Vector3(0, 0);
-            uncheckSprite.spriteName = "check-unchecked";
-
-            UISprite checkSprite = cb.AddUIComponent<UISprite>();
-            checkSprite.height = 20;
-            checkSprite.width = 20;
-            checkSprite.relativePosition = new Vector3(0, 0);
-            checkSprite.spriteName = "check-checked";
-
-            cb.checkedBoxObject = checkSprite;
-            cb.disabledColor = new Color(127.0f / 255.0f, 127.0f / 255.0f, 127.0f / 255.0f, 1.0f);
-            label.disabledColor = new Color(127.0f / 255.0f, 127.0f / 255.0f, 127.0f / 255.0f, 1.0f);
-            cb.tooltip = tooltip;
-            label.tooltip = cb.tooltip;
-            cb.isChecked = false;
-            cb.label.isVisible = hidden;
-            cb.isVisible = hidden;
-
-            //WriteLog("Leaving addCheckbox");
-            return cb;
-        }
-        
-        private UIDropDown addDropDown(UIPanel panel, int y, int x, int w, int h, string text, string tooltip)
-        {
-            UIDropDown dd = panel.AddUIComponent<UIDropDown>();
-
-            dd.size = new Vector2(w, h);
-            dd.relativePosition = new Vector3(x, y);
-            dd.listBackground = "GenericPanelLight";
-            dd.itemHeight = 15;
-            dd.itemHover = "ListItemHover";
-            dd.itemHighlight = "ListItemHighlight";
-            dd.normalBgSprite = "ListItemHover";
-            dd.listWidth = 100;
-            dd.listHeight = 350;
-            dd.foregroundSpriteMode = UIForegroundSpriteMode.Stretch;
-            dd.popupColor = new Color32(45, 52, 61, 255);
-            dd.popupTextColor = new Color32(170, 170, 170, 255);
-            dd.zOrder = 1;
-            dd.textScale = 0.8f;
-            dd.verticalAlignment = UIVerticalAlignment.Middle;
-            dd.horizontalAlignment = UIHorizontalAlignment.Left;
-            dd.selectedIndex = 0;
-            dd.textFieldPadding = new RectOffset(8, 0, 8, 0);
-            dd.itemPadding = new RectOffset(14, 0, 0, 0);
-            
-            var dropdownButton = dd.AddUIComponent<UIButton>();
-            dd.triggerButton = dropdownButton;
-
-            dropdownButton.text = "";
-            dropdownButton.size = dd.size;
-            dropdownButton.relativePosition = new Vector3(0.0f, 0.0f);
-            dropdownButton.textVerticalAlignment = UIVerticalAlignment.Middle;
-            dropdownButton.textHorizontalAlignment = UIHorizontalAlignment.Left;
-            dropdownButton.normalFgSprite = "IconDownArrow";
-            dropdownButton.hoveredFgSprite = "IconDownArrowHovered";
-            dropdownButton.pressedFgSprite = "IconDownArrowPressed";
-            dropdownButton.focusedFgSprite = "IconDownArrowFocused";
-            dropdownButton.disabledFgSprite = "IconDownArrowDisabled";
-            dropdownButton.foregroundSpriteMode = UIForegroundSpriteMode.Fill;
-            dropdownButton.horizontalAlignment = UIHorizontalAlignment.Right;
-            dropdownButton.verticalAlignment = UIVerticalAlignment.Middle;
-            dropdownButton.zOrder = 0;
-            dropdownButton.textScale = 0.8f;
-            return dd;
-        }
-
-        private UISlider addSlider(UIPanel panel, string name, int y, int x, int w, int h, float min, float max, float step, float defaultValue, string tooltip)
-        {
-            UISlider sl = panel.AddUIComponent<UISlider>();
-
-            sl.relativePosition = new Vector3(x, y);
-            sl.name = name;
-            sl.width = w;
-            sl.height = h;
-            sl.tooltip = tooltip;
-            sl.minValue = min;
-            sl.maxValue = max;
-            sl.stepSize = step;
-            sl.value = defaultValue;
-            sl.isVisible = true;
-            sl.color = Color.blue;
-            sl.BringToFront();
-            if (mode == LoadMode.LoadMap || mode == LoadMode.NewMap)
-                sl.backgroundSprite = "SubcategoriesPanel";
-            else
-                sl.backgroundSprite = "GenericPanel";
-
-            return sl;
-        }
-
-        private UITextField addTextBox(UIPanel panel, string name, string text, int y, int x, int width, int height, string tooltip, bool numeric, bool allowFloats)
-        {
-            UITextField tf = panel.AddUIComponent<UITextField>();
-
-            tf.relativePosition = new Vector3(x, y);
-            tf.size = new Vector3(width, height);
-            tf.name = name;
-            tf.text = text;
-            tf.width = width;
-            tf.height = height;
-            tf.tooltip = tooltip;
-            tf.numericalOnly = numeric;
-            tf.allowFloats = allowFloats;
-            tf.textScale = 0.8f;
-            tf.color = Color.black;
-            tf.cursorBlinkTime = 0.45f;
-            tf.cursorWidth = 1;
-            tf.horizontalAlignment = UIHorizontalAlignment.Left;
-            tf.selectionBackgroundColor = new Color(233, 201, 148, 255);
-            tf.selectionSprite = "EmptySprite";
-            tf.verticalAlignment = UIVerticalAlignment.Middle;
-            tf.padding = new RectOffset(5, 0, 5, 0);
-            tf.foregroundSpriteMode = UIForegroundSpriteMode.Fill;
-            tf.normalBgSprite = "TextFieldPanel";
-            tf.hoveredBgSprite = "TextFieldPanelHovered";
-            tf.focusedBgSprite = "TextFieldPanel";
-            tf.isInteractive = true;
-            tf.enabled = true;
-            tf.readOnly = false;
-            tf.builtinKeyNavigation = true;
-
-            return tf;
-        }
-        
-        private UIButton addButton(UIPanel panel, string text, string tooltip, int y, int x, int w, int h)
-        {
-            UIButton bt = panel.AddUIComponent<UIButton>();
-            bt.relativePosition = new Vector3(x, y);
-            bt.name = text.Replace(" ", "_");
-            bt.text = text;
-            bt.tooltip = tooltip;
-            bt.textScale = 0.8f;
-            bt.width = w;
-            bt.height = h;
-            bt.normalBgSprite = "ButtonMenu";
-            bt.disabledBgSprite = "ButtonMenuDisabled";
-            bt.hoveredBgSprite = "ButtonMenuHovered";
-            bt.focusedBgSprite = "ButtonMenu";
-            bt.pressedBgSprite = "ButtonMenuPressed";
-            bt.textColor = new Color32(255, 255, 255, 255);
-            bt.disabledTextColor = new Color32(7, 7, 7, 255);
-            bt.hoveredTextColor = new Color32(255, 255, 255, 255);
-            bt.focusedTextColor = new Color32(255, 255, 255, 255);
-            bt.pressedTextColor = new Color32(30, 30, 44, 255);
-
-            //WriteLog("Leaving addButton: " + text + " location: " + bt.relativePosition + "main HeightxWidth: " + panel.height + "x" + panel.width);
-            return bt;
-        }
-
-        #endregion
-
-        #endregion
 
         #region "Settings"
 
@@ -1204,65 +351,65 @@ namespace AnotherRoadUpdateTool
             if (m_settings == true) { return; }
             m_settings = true;
 
-            //WriteLog("Leaving SetSettings types[(int)tp.Ground].isChecked: " + types[(int)tp.Ground].isChecked);
+            //WriteLog("Leaving SetSettings plMain.Types(PanelMain.tp.Ground): " + plMain.Types(PanelMain.tp.Ground));
 
-            us.ToOneway = toTypes[(int)Up.Oneway].isChecked;
-            us.ToMedium = toTypes[(int)Up.Medium].isChecked;
-            us.ToHighway = toTypes[(int)Up.Highway].isChecked;
-            us.ToLarge = toTypes[(int)Up.Large].isChecked;
-            us.ToBasic = toTypes[(int)Up.Basic].isChecked;
-            us.Oneway = fromTypes[(int)Up.Oneway].isChecked;
-            us.Medium = fromTypes[(int)Up.Medium].isChecked;
-            us.Highway = fromTypes[(int)Up.Highway].isChecked;
-            us.Large = fromTypes[(int)Up.Large].isChecked;
-            us.Basic = fromTypes[(int)Up.Basic].isChecked;
+            us.ToOneway = plMain.To(PanelMain.up.Oneway);
+            us.ToMedium = plMain.To(PanelMain.up.Medium);
+            us.ToHighway = plMain.To(PanelMain.up.Highway);
+            us.ToLarge = plMain.To(PanelMain.up.Large);
+            us.ToBasic = plMain.To(PanelMain.up.Basic);
+            us.Oneway = plMain.From(PanelMain.up.Oneway);
+            us.Medium = plMain.From(PanelMain.up.Medium);
+            us.Highway = plMain.From(PanelMain.up.Highway);
+            us.Large = plMain.From(PanelMain.up.Large);
+            us.Basic = plMain.From(PanelMain.up.Basic);
 
-            us.Roads = deletes[(int)p.Roads].isChecked;
-            us.Railroads = deletes[(int)p.Railroads].isChecked;
-            us.Highways = deletes[(int)p.Highways].isChecked;
-            us.PowerLines = deletes[(int)p.PowerLines].isChecked;
-            us.WaterPipes = deletes[(int)p.WaterPipes].isChecked;
-            us.HeatPipes = deletes[(int)p.HeatPipes].isChecked;
-            us.Airplanes = deletes[(int)p.Airplanes].isChecked;
-            us.Shipping = deletes[(int)p.Shipping].isChecked;
-            us.Pedestrian = deletes[(int)p.Pedestrian].isChecked;
-            us.Bicycle = deletes[(int)p.Bicycle].isChecked;
-            us.Tram = deletes[(int)p.Tram].isChecked;
-            us.Metro = deletes[(int)p.Metro].isChecked;
+            us.Roads = plMain.Deletes(PanelMain.dl.Roads);
+            us.Railroads = plMain.Deletes(PanelMain.dl.Railroads);
+            us.Highways = plMain.Deletes(PanelMain.dl.Highways);
+            us.PowerLines = plMain.Deletes(PanelMain.dl.PowerLines);
+            us.WaterPipes = plMain.Deletes(PanelMain.dl.WaterPipes);
+            us.HeatPipes = plMain.Deletes(PanelMain.dl.HeatPipes);
+            us.Airplanes = plMain.Deletes(PanelMain.dl.Airplanes);
+            us.Shipping = plMain.Deletes(PanelMain.dl.Shipping);
+            us.Pedestrian = plMain.Deletes(PanelMain.dl.Pedestrian);
+            us.Bicycle = plMain.Deletes(PanelMain.dl.Bicycle);
+            us.Tram = plMain.Deletes(PanelMain.dl.Tram);
+            us.Metro = plMain.Deletes(PanelMain.dl.Metro);
 
-            us.Ground = types[(int)tp.Ground].isChecked;
-            us.Bridge = types[(int)tp.Bridge].isChecked;
-            us.Slope = types[(int)tp.Slope].isChecked;
-            us.Tunnel = types[(int)tp.Tunnel].isChecked;
-            us.Curve = types[(int)tp.Curve].isChecked;
+            us.Ground = plMain.Types(PanelMain.tp.Ground);
+            us.Bridge = plMain.Types(PanelMain.tp.Bridge);
+            us.Slope = plMain.Types(PanelMain.tp.Slope);
+            us.Tunnel = plMain.Types(PanelMain.tp.Tunnel);
+            us.Curve = plMain.Types(PanelMain.tp.Curve);
 
-            us.Buildings = deletes[(int)p.Buildings].isChecked;
-            us.Trees = deletes[(int)p.Trees].isChecked;
-            us.Props = deletes[(int)p.Props].isChecked;
+            us.Buildings = plMain.Deletes(PanelMain.dl.Buildings);
+            us.Trees = plMain.Deletes(PanelMain.dl.Trees);
+            us.Props = plMain.Deletes(PanelMain.dl.Props);
 
-            us.Update = options[(int)ops.Updates].isChecked;
-            us.Delete = options[(int)ops.Deletes].isChecked;
-            us.Services = options[(int)ops.Services].isChecked;
-            us.Terrain = options[(int)ops.Terrain].isChecked;
-            us.Toggle = cbToggle.isChecked;
-            us.Districts = options[(int)ops.Districts].isChecked;
-            us.DistrictToggle = cbDistrictToggle.isChecked;
-            us.TerrainHeight = m_terrainHeight;
+            us.Update = plMain.Options(PanelMain.ops.Updates);
+            us.Delete = plMain.Options(PanelMain.ops.Deletes);
+            us.Services = plMain.Options(PanelMain.ops.Services);
+            us.Terrain = plMain.Options(PanelMain.ops.Terrain);
+            us.Toggle = plMain.cbToggle.isChecked;
+            us.Districts = plMain.Options(PanelMain.ops.Districts);
+            us.DistrictToggle = plMain.cbDistrictToggle.isChecked;
+            us.TerrainHeight = plMain.TerrainHeight;
 
-            us.HealthCare = services[(int)dl.HealthCare].isChecked;
-            us.PoliceDepartment = services[(int)dl.PoliceDepartment].isChecked;
-            us.FireDepartment = services[(int)dl.FireDepartment].isChecked;
-            us.PublicTransport = services[(int)dl.PublicTransport].isChecked;
-            us.Education = services[(int)dl.Education].isChecked;
-            us.Electricity = services[(int)dl.Electricity].isChecked;
-            us.Water = services[(int)dl.Water].isChecked;
-            us.Garbage = services[(int)dl.Garbage].isChecked;
-            us.Beautification = services[(int)dl.Beautification].isChecked;
-            us.Monument = services[(int)dl.Monument].isChecked;
-            us.Abandoned = services[(int)dl.Abandoned].isChecked;
-            us.Burned = services[(int)dl.Burned].isChecked;
+            us.HealthCare = plMain.Services(PanelMain.sr.HealthCare);
+            us.PoliceDepartment = plMain.Services(PanelMain.sr.PoliceDepartment);
+            us.FireDepartment = plMain.Services(PanelMain.sr.FireDepartment);
+            us.PublicTransport = plMain.Services(PanelMain.sr.PublicTransport);
+            us.Education = plMain.Services(PanelMain.sr.Education);
+            us.Electricity = plMain.Services(PanelMain.sr.Electricity);
+            us.Water = plMain.Services(PanelMain.sr.Water);
+            us.Garbage = plMain.Services(PanelMain.sr.Garbage);
+            us.Beautification = plMain.Services(PanelMain.sr.Beautification);
+            us.Monument = plMain.Services(PanelMain.sr.Monument);
+            us.Abandoned = plMain.Services(PanelMain.sr.Abandoned);
+            us.Burned = plMain.Services(PanelMain.sr.Burned);
 
-            us.Chirper = services[(int)dl.Chirper].isChecked;
+            us.Chirper = plMain.Services(PanelMain.sr.Chirper);
             us.ShowDelete = ShowDelete;
             us.ShowUpdate = ShowUpdate;
             us.ShowTerrain = ShowTerrain;
@@ -1280,11 +427,9 @@ namespace AnotherRoadUpdateTool
 
             us.MaxAreas = MaxAreas;
             us.StartMoney = StartMoney;
-            //WriteLog("On Exit us.Abandoned & us.Burned are: " + us.Abandoned + " & " + us.Burned);
 
             us.Save();
-
-            //WriteLog("Leaving SetSettings types[(int)tp.Ground].isChecked: " + types[(int)tp.Ground].isChecked);
+            
             m_settings = false;
         }
 
@@ -1310,39 +455,39 @@ namespace AnotherRoadUpdateTool
 
             try
             {
-                toTypes[(int)Up.Oneway].isChecked = us.ToOneway;
-                toTypes[(int)Up.Medium].isChecked = us.ToMedium;
-                toTypes[(int)Up.Highway].isChecked = us.ToHighway;
-                toTypes[(int)Up.Large].isChecked = us.ToLarge;
-                toTypes[(int)Up.Basic].isChecked = us.ToBasic;
-                fromTypes[(int)Up.Oneway].isChecked = us.Oneway;
-                fromTypes[(int)Up.Medium].isChecked = us.Medium;
-                fromTypes[(int)Up.Highway].isChecked = us.Highway;
-                fromTypes[(int)Up.Large].isChecked = us.Large;
-                fromTypes[(int)Up.Basic].isChecked = us.Basic;
+                plMain.To(PanelMain.up.Oneway, true, us.ToOneway);
+                plMain.To(PanelMain.up.Medium, true, us.ToMedium);
+                plMain.To(PanelMain.up.Highway, true, us.ToHighway);
+                plMain.To(PanelMain.up.Large, true, us.ToLarge);
+                plMain.To(PanelMain.up.Basic, true, us.ToBasic);
+                plMain.From(PanelMain.up.Oneway, true, us.Oneway);
+                plMain.From(PanelMain.up.Medium, true, us.Medium);
+                plMain.From(PanelMain.up.Highway, true, us.Highway);
+                plMain.From(PanelMain.up.Large, true, us.Large);
+                plMain.From(PanelMain.up.Basic, true, us.Basic);
 
-                deletes[(int)p.Roads].isChecked = us.Roads;
-                deletes[(int)p.Railroads].isChecked = us.Railroads;
-                deletes[(int)p.Highways].isChecked = us.Highways;
-                deletes[(int)p.PowerLines].isChecked = us.PowerLines;
-                deletes[(int)p.WaterPipes].isChecked = us.WaterPipes;
-                deletes[(int)p.HeatPipes].isChecked = us.HeatPipes;
-                deletes[(int)p.Airplanes].isChecked = us.Airplanes;
-                deletes[(int)p.Shipping].isChecked = us.Shipping;
-                deletes[(int)p.Pedestrian].isChecked = us.Pedestrian;
-                deletes[(int)p.Bicycle].isChecked = us.Bicycle;
-                deletes[(int)p.Tram].isChecked = us.Tram;
-                deletes[(int)p.Metro].isChecked = us.Metro;
+                plMain.Deletes(PanelMain.dl.Roads, true, us.Roads);
+                plMain.Deletes(PanelMain.dl.Railroads, true, us.Railroads);
+                plMain.Deletes(PanelMain.dl.Highways, true, us.Highways);
+                plMain.Deletes(PanelMain.dl.PowerLines, true, us.PowerLines);
+                plMain.Deletes(PanelMain.dl.WaterPipes, true, us.WaterPipes);
+                plMain.Deletes(PanelMain.dl.HeatPipes, true, us.HeatPipes);
+                plMain.Deletes(PanelMain.dl.Airplanes, true, us.Airplanes);
+                plMain.Deletes(PanelMain.dl.Shipping, true, us.Shipping);
+                plMain.Deletes(PanelMain.dl.Pedestrian, true, us.Pedestrian);
+                plMain.Deletes(PanelMain.dl.Bicycle, true, us.Bicycle);
+                plMain.Deletes(PanelMain.dl.Tram, true, us.Tram);
+                plMain.Deletes(PanelMain.dl.Metro, true, us.Metro);
 
-                types[(int)tp.Ground].isChecked = us.Ground;
-                types[(int)tp.Bridge].isChecked = us.Bridge;
-                types[(int)tp.Slope].isChecked = us.Slope;
-                types[(int)tp.Tunnel].isChecked = us.Tunnel;
-                types[(int)tp.Curve].isChecked = us.Curve;
+                plMain.Types(PanelMain.tp.Ground, true, us.Ground);
+                plMain.Types(PanelMain.tp.Bridge, true, us.Bridge);
+                plMain.Types(PanelMain.tp.Slope, true, us.Slope);
+                plMain.Types(PanelMain.tp.Tunnel, true, us.Tunnel);
+                plMain.Types(PanelMain.tp.Curve, true, us.Curve);
 
-                deletes[(int)p.Buildings].isChecked = us.Buildings;
-                deletes[(int)p.Trees].isChecked = us.Trees;
-                deletes[(int)p.Props].isChecked = us.Props;
+                plMain.Deletes(PanelMain.dl.Buildings, true, us.Buildings);
+                plMain.Deletes(PanelMain.dl.Trees, true, us.Trees);
+                plMain.Deletes(PanelMain.dl.Props, true, us.Props);
 
                 if (mode == LoadMode.LoadMap || mode == LoadMode.NewMap)
                 {
@@ -1362,36 +507,34 @@ namespace AnotherRoadUpdateTool
                 }
 
                 loc  = "Services";
-                options[(int)ops.Services].isChecked = us.Services;
+                plMain.Options(PanelMain.ops.Services, true, us.Services);
                 loc = "Deletes";
-                options[(int)ops.Deletes].isChecked = us.Delete;
+                plMain.Options(PanelMain.ops.Deletes, true, us.Delete);
                 loc = "Updates";
-                options[(int)ops.Updates].isChecked = us.Update;
+                plMain.Options(PanelMain.ops.Updates, true, us.Update);
                 loc = "Terrain";
-                options[(int)ops.Terrain].isChecked = us.Terrain;
+                plMain.Options(PanelMain.ops.Terrain, true, us.Terrain);
                 loc = "Toggle";
-                cbToggle.isChecked = us.Toggle;
-                options[(int)ops.Districts].isChecked = us.Districts;
-                cbDistrictToggle.isChecked = us.DistrictToggle;
+                plMain.cbToggle.isChecked = us.Toggle;
+                plMain.Options(PanelMain.ops.Districts, true, us.Districts);
+                plMain.cbDistrictToggle.isChecked = us.DistrictToggle;
                 loc = "TerrainHeight";
-                m_terrainHeight = (float)us.TerrainHeight;
-                lHeight.text = m_terrainHeight.ToString("0.00");
-                tfTerrainHeight.text = lHeight.text;
+                plMain.TerrainHeight = us.TerrainHeight;
 
-                services[(int)dl.HealthCare].isChecked = us.HealthCare;
-                services[(int)dl.PoliceDepartment].isChecked = us.PoliceDepartment;
-                services[(int)dl.FireDepartment].isChecked = us.FireDepartment;
-                services[(int)dl.PublicTransport].isChecked = us.PublicTransport;
-                services[(int)dl.Education].isChecked = us.Education;
-                services[(int)dl.Electricity].isChecked = us.Electricity;
-                services[(int)dl.Water].isChecked = us.Water;
-                services[(int)dl.Garbage].isChecked = us.Garbage;
-                services[(int)dl.Beautification].isChecked = us.Beautification;
-                services[(int)dl.Monument].isChecked = us.Monument;
-                services[(int)dl.Abandoned].isChecked = us.Abandoned;
-                services[(int)dl.Burned].isChecked = us.Burned;
+                plMain.Services(PanelMain.sr.HealthCare, true, us.HealthCare);
+                plMain.Services(PanelMain.sr.PoliceDepartment, true, us.PoliceDepartment);
+                plMain.Services(PanelMain.sr.FireDepartment, true, us.FireDepartment);
+                plMain.Services(PanelMain.sr.PublicTransport, true, us.PublicTransport);
+                plMain.Services(PanelMain.sr.Education, true, us.Education);
+                plMain.Services(PanelMain.sr.Electricity, true, us.Electricity);
+                plMain.Services(PanelMain.sr.Water, true, us.Water);
+                plMain.Services(PanelMain.sr.Garbage, true, us.Garbage);
+                plMain.Services(PanelMain.sr.Beautification, true, us.Beautification);
+                plMain.Services(PanelMain.sr.Monument, true, us.Monument);
+                plMain.Services(PanelMain.sr.Abandoned, true, us.Abandoned);
+                plMain.Services(PanelMain.sr.Burned, true, us.Burned);
                 loc = "Chirper";
-                services[(int)dl.Chirper].isChecked = us.Chirper;
+                plMain.Services(PanelMain.sr.Chirper, true, us.Chirper);
 
                 ShowDelete = us.ShowDelete;
                 ShowUpdate = us.ShowUpdate;
@@ -1416,819 +559,25 @@ namespace AnotherRoadUpdateTool
                 //we need to toggle shown or not
                 if (mode != LoadMode.LoadMap && mode != LoadMode.NewMap)
                 {
-                    Chirp.Toggle(services[(int)dl.Chirper].isChecked);
+                    Chirp.Toggle(plMain.Services(PanelMain.sr.Chirper));
                 }
                 loc = "ZoneTool";
             }
             catch (Exception ex)
             {
-                WriteError("Error in GetSettings loc: " + loc + " deltypes.Count: " + deletes.Count, ex);
+                WriteError("Error in GetSettings loc: " + loc + " deltypes.Count: " + plMain.deletes.Count, ex);
             }
-            //WriteLog("Leaving GetSettings types[(int)tp.Ground].isChecked: " + types[(int)tp.Ground].isChecked);
+            //WriteLog("Leaving GetSettings plMain.Types(PanelMain.tp.Ground): " + plMain.Types(PanelMain.tp.Ground));
             m_settings = false;
         }
                        
         #endregion
 
-        #region "Event Process helpers"
-
-        private void SetServicesEnabled()
-        {
-            //WriteLog("Entering: SetServicesEnabled m_selectable: " + m_selectable);
-            try
-            {
-                m_selectable = false;
-                lSelectable.text = m_updatetool + m_unavailable;
-
-                //do we have at least one service type selected
-                if (services.Any(o => o.isChecked == true)) { }
-                else
-                    return;
-
-                //ok all checks complete
-                m_selectable = true;
-                lSelectable.text = m_updatetool + m_available;
-            }
-            catch (Exception ex)
-            {
-                WriteError("Error in SetServicesEnabled Exception: ", ex);
-            }
-            //WriteLog("Leaving: SetServicesEnabled m_selectable: " + m_selectable);
-        }
-
-        private void SetTerrainEnabled()
-        {
-            if (lHeight.text == tfTerrainHeight.text)
-            {
-                m_selectable = true;
-                lSelectable.text = m_updatetool + m_available;
-                lInformation.text = m_defaultInfo;
-            }
-            else
-            {
-                m_selectable = false;
-                lSelectable.text = m_updatetool + m_unavailable;
-                lInformation.text = "Use the Validate button to validate your input";
-            }           
-        }
-
-        private void SetDistrictsEnabled()
-        {
-            //WriteLog("Entering: SetDistrictsEnabled m_selectable: " + m_selectable);
-            try
-            {
-                m_selectable = true;
-                lSelectable.text = m_updatetool + m_available;
-            }
-            catch (Exception ex)
-            {
-                WriteError("Error in SetDistrictsEnabled Exception: ", ex);
-            }
-            //WriteLog("Leaving: SetDistrictsEnabled m_selectable: " + m_selectable);
-        }
-
-        private void SetUpdateEnabled()
-        {
-            //WriteLog("Entering: SetUpdateEnabled m_selectable: " + m_selectable);
-
-            try
-            {
-                m_selectable = false;
-                lSelectable.text = m_updatetool + m_unavailable;
-
-                //is Updated enabled?
-                if (options[(int)ops.Updates].isChecked == false)
-                {
-                    //WriteLog("Leaving: SetUpdateEnabled 'Update not checked' m_selectable: " + m_selectable);
-                    return;
-                }
-
-                //do we have a to road type
-                if (types.Any(o => o.isChecked == true)) { }
-                else
-                {
-                    //WriteLog("Leaving: SetUpdateEnabled ' No types selected' m_selectable: " + m_selectable);
-                    return;
-                }
-
-                //do we have a from road type
-                if (fromTypes.Any(o => o.isChecked == true)) { }
-                else
-                {
-                    //WriteLog("Leaving: SetUpdateEnabled 'No to road type selected' m_selectable: " + m_selectable);
-                    return;
-                }
-
-                //do we have a to road type
-                if (toTypes.Any(o => o.isChecked == true)) { }
-                else
-                {
-                    //WriteLog("Leaving: SetUpdateEnabled 'No from road type selected' m_selectable: " + m_selectable);
-                    return;
-                }
-
-                //do we have a from Road
-                if (fromOneway.Any(o => o.isChecked == true)) { }
-                else if (fromMedium.Any(o => o.isChecked == true)) { }
-                else if (fromHighway.Any(o => o.isChecked == true)) { }
-                else if (fromLarge.Any(o => o.isChecked == true)) { }
-                else if (fromBasic.Any(o => o.isChecked == true)) { }
-                else
-                {
-                    //WriteLog("Leaving: SetUpdateEnabled 'No to road selected'm_selectable: " + m_selectable);
-                    return;
-                }
-
-                //do we have a to Road
-                if (toOneway.Any(o => o.isChecked == true)) { }
-                else if (toMedium.Any(o => o.isChecked == true)) { }
-                else if (toHighway.Any(o => o.isChecked == true)) { }
-                else if (toLarge.Any(o => o.isChecked == true)) { }
-                else if (toBasic.Any(o => o.isChecked == true)) { }
-                else
-                {
-                    //WriteLog("Leaving: SetUpdateEnabled 'No from road selected' m_selectable: " + m_selectable);
-                    return;
-                }
-
-                //ok all checks complete
-                m_selectable = true;
-                lSelectable.text = m_updatetool + m_available;
-            }
-            catch (Exception ex)
-            {
-                WriteError("Error in Error in SetUpdateEnabled ", ex);
-            }
-
-            //WriteLog("Leaving: SetUpdateEnabled m_selectable: " + m_selectable);
-        }
-
-        private void SetDeleteEnabled()
-        {
-            //WriteLog("Entering SetDeleteEnabled selectable is: " + m_selectable);
-            try
-            {                
-                m_selectable = false;
-                lSelectable.text = m_updatetool + m_unavailable;
-
-                //is delete enabled?
-                if (options[(int)ops.Deletes].isChecked == false)
-                {
-                    //WriteLog("Leaving: SetDeleteEnabled 'Deletes not checked' m_selectable: " + m_selectable);
-                    return;
-                }
-
-                //do we have any types
-                if (types.Any(o => o.isChecked == true)) { }
-                else
-                {
-                    //WriteLog("Leaving: SetDeleteEnabled 'No types selected' m_selectable: " + m_selectable);
-                    return;
-                }
-                //do we have anyhting checked to delete?
-                if (deletes.Any(o => o.isChecked == true)) { }
-                else
-                {
-                    //WriteLog("Leaving: SetDeleteEnabled 'No item selected to delete' m_selectable: " + m_selectable);
-                    return;
-                }
-                m_selectable = true;
-                lSelectable.text = m_updatetool + m_available;
-            }
-            catch (Exception ex)
-            {
-                WriteError("Error in SetDeleteEnabled Exception: ", ex);
-            }
-            //WriteLog("Leaving SetDeleteEnabled selectable is: " + m_selectable);
-        }
-
-        private void ShowRoads()
-        {
-            //From
-            //Make sure all road panels are hidden
-            plBasic.isVisible = false;
-            plHighway.isVisible = false;
-            plLarge.isVisible = false;
-            plMedium.isVisible = false;
-            plOneway.isVisible = false;
-            //now check for any that should be displayed
-            foreach (UICheckBox c in fromTypes)
-            {
-                if (c.isChecked == true && options[(int)ops.Updates].isChecked)
-                {
-                    if (c.text == "Basic") { plBasic.isVisible = true; }
-                    if (c.text == "Highway") { plHighway.isVisible = true; }
-                    if (c.text == "Large") { plLarge.isVisible = true; }
-                    if (c.text == "Medium") { plMedium.isVisible = true; }
-                    if (c.text == "Oneway") { plOneway.isVisible = true; }
-                }
-            }
-            //To
-            //Make sure all road panels are hidden
-            plToBasic.isVisible = false;
-            plToHighway.isVisible = false;
-            plToLarge.isVisible = false;
-            plToMedium.isVisible = false;
-            plToOneway.isVisible = false;
-            //now check for any that should be displayed
-            foreach (UICheckBox c in toTypes)
-            {
-                if (c.isChecked == true && options[(int)ops.Updates].isChecked)
-                {
-                    if (c.text == "Basic") { plToBasic.isVisible = true; }
-                    if (c.text == "Highway") { plToHighway.isVisible = true; }
-                    if (c.text == "Large") { plToLarge.isVisible = true; }
-                    if (c.text == "Medium") { plToMedium.isVisible = true; }
-                    if (c.text == "Oneway") { plToOneway.isVisible = true; }
-                }
-            }
-            //Show if available
-            SetUpdateEnabled();
-        }
-
-        private void UpdatePanels(bool show)
-        {
-            plRoads.isVisible = show;
-            plDelete.isVisible = show;
-            plServices.isVisible = show;
-            plTerrain.isVisible = show;
-            plBasic.isVisible = show;
-            plHighway.isVisible = show;
-            plLarge.isVisible = show;
-            plMedium.isVisible = show;
-            plOneway.isVisible = show;
-            plToBasic.isVisible = show;
-            plToHighway.isVisible = show;
-            plToLarge.isVisible = show;
-            plToMedium.isVisible = show;
-            plToOneway.isVisible = show;
-            plDistricts.isVisible = show;
-
-            SetUpdateEnabled();
-        }
-
-        private void UpdateDisplayedRoads(List<UICheckBox> types, List<UICheckBox> roads, string text, bool show, int xPos)
-        {
-            //WriteLog("Entering UpdateDisplayedRoads " + xPos);
-            foreach (UICheckBox cb in types)
-            {
-                if (cb.text != text)
-                {
-                    cb.isChecked = false;
-                }
-            }
-
-            if (text.Contains("Basic")) { DisplayCheckBoxes(roads, xPos, "Basic", show); }
-            else if (text.Contains("Highway")) { DisplayCheckBoxes(roads, xPos, "Highway", show); }
-            else if (text.Contains("Large")) { DisplayCheckBoxes(roads, xPos, "Large", show); }
-            else if (text.Contains("Medium")) { DisplayCheckBoxes(roads, xPos, "Medium", show); }
-            else if (text.Contains("Oneway")) { DisplayCheckBoxes(roads, xPos, "Oneway", show); }
-            else
-                WriteLog("Could not define this object type.");
-            //WriteLog("Leaving UpdateDisplayedRoads");
-        }
-
-        private void DisplayCheckBoxes(List<UICheckBox> roads, int xPos, string test, bool show)
-        {
-            //WriteLog("Entering DisplayCheckBoxes");
-
-            int y = 0;
-            int x = xPos;
-
-            foreach (UICheckBox cb in roads)
-            {
-                cb.isVisible = false;
-                cb.label.isVisible = false;
-                if (cb.text.Contains(test))
-                {
-                    cb.relativePosition = new Vector3(x, y);
-                    cb.label.relativePosition = new Vector3(x + 25, y + 3);
-                    cb.isVisible = show;
-                    cb.label.isVisible = show;
-                    y += 25;
-                }
-            }
-
-            //WriteLog("Leaving UpdateDisplayedRoads");
-        }
-
-        private int ReturnInteger(string value, int min, int max)
-        {
-            int tmp = min;
-            try
-            {
-                tmp = Convert.ToInt32(value);
-            }
-            catch (Exception)
-            {
-                tmp = 0;
-            }
-            return Math.Min(Math.Max(tmp, min), max);
-        }
-
-        private double ReturnDouble(string value, double min, double max)
-        {
-            double tmp = min;
-            try
-            {
-                tmp = Convert.ToDouble(value);
-            }
-            catch (Exception)
-            {
-                tmp = 0;
-            }
-            return Math.Min(Math.Max(tmp, min), max);
-        }
-
-        private bool isNumeric(string value)
-        {
-            bool results = false;
-            try
-            {
-                System.Convert.ToInt32(value);
-                results = true;
-            }
-            catch (FormatException e)
-            {
-                WriteError("Error in isNumeric: ", e);
-            }
-            return results;
-        }
-
-        #endregion
-
         #region "Event Handlers"
-
-
-        private void Main_eventSizeChanged(UIComponent component, Vector2 value)
-        {
-            if (us.Width < 250)
-                us.Width = 250;
-            if (us.Height < 400)
-                us.Height = 400;
-            //set the values
-            if (plMain.width > us.Width)
-                plMain.width = us.Width;
-            if (plMain.height > us.Height)
-                plMain.height = us.Height;
-            lbTitle.width = plMain.width;
-        }
-
-        private void Main_eventMouseUp(UIComponent component, UIMouseEventParameter eventParam)
-        {
-            if (mouseDown == true)
-            {
-                mouseDown = false;
-                MoveCompleted();
-            }
-        }
-
-        private void Main_eventMouseMove(UIComponent component, UIMouseEventParameter eventParam)
-        {
-            if (mouseDown == true)
-            {
-                try
-                {
-                    // Move the top and left according to the delta amount
-                    Vector3 delta = new Vector3(eventParam.moveDelta.x, eventParam.moveDelta.y);
-                    //Just move the Panel
-                    plMain.position += delta;
-                    PanelPosition.left = (int)delta.x;
-                    PanelPosition.top = (int)delta.y;
-                }
-                catch (Exception ex)
-                {
-                    WriteError("Error in MouseMove: ", ex);
-                }
-            }
-        }
-
-        private void Main_eventMouseDown(UIComponent component, UIMouseEventParameter eventParam)
-        {
-            mouseDown = true;
-        }
-
-        private void MoveCompleted()
-        {
-            us.Height = PanelPosition.height;
-            us.Left = PanelPosition.left;
-            us.Top = PanelPosition.top;
-            us.Width = PanelPosition.width;
-        }
-
-        private void GameHeight_eventClick(UIComponent component, UIMouseEventParameter eventParam)
-        {
-            string found = "TSBar";
-            try
-            {
-                UISlicedSprite TSBar = (UISlicedSprite)UIView.GetAView().FindUIComponent("TSBar");
-                found = "OptionsBar";
-                UIPanel OptionsBar = (UIPanel)TSBar.Find("OptionsBar");
-                found = "LevelHeight";
-                UIPanel LevelHeight = (UIPanel)OptionsBar.Find("LevelHeightPanel");
-                found = "Settings";
-                UIPanel Settings = (UIPanel)LevelHeight.Find("Settings");
-                found = "Height";
-                UISlider Height = (UISlider)Settings.Find("Height");
-                found = "Parsing " + Height.value;
-
-                tfTerrainHeight.text = Height.value.ToString("0.00");
-            }
-            catch (Exception ex)
-            {
-                ARUT.WriteError("Object Not found in MyITerrain.TerrainHeight. " + found, ex);
-            }
-        }
-
-        private void btHide_eventClick(UIComponent component, UIMouseEventParameter eventParam)
-        {
-            //WriteLog("Entering btHide_eventClick: " + plMain.isVisible);
-            this.OnDisable();
-            plMain.isVisible = false;
-            this.enabled = false;
-            //WriteLog("Leaving btHide_eventClick: " + plMain.isVisible);
-        }
-
-        private void TerrainHeight_eventTextChanged(UIComponent component, string value)
-        {
-            SetTerrainEnabled();
-        }
-
-        private void TerrainHeight_eventKeyDown(UIComponent component, UIKeyEventParameter eventParam)
-        {
-            if ((eventParam.keycode >= KeyCode.Alpha0 && eventParam.keycode <= KeyCode.Alpha9) || eventParam.keycode <= KeyCode.Period) { }
-            else
-            { eventParam = null;  }
-        }
-
-        private void Validate_eventClick(UIComponent component, UIMouseEventParameter eventParam)
-        {
-            if (btValidate.state == UIButton.ButtonState.Disabled) { return; }
-
-            string raw = tfTerrainHeight.text;
-            double val;
-            if (double.TryParse(raw, out val) == true)
-            {
-                if (val > 2000)
-                {
-                    tfTerrainHeight.text = "2000.00";
-                }
-                else if (val < 0)
-                {
-                    tfTerrainHeight.text = "0.00";
-                }
-                else
-                {
-                    tfTerrainHeight.text = val.ToString("0.00");
-                }
-                m_terrainHeight = (float)val;
-                lHeight.text = val.ToString("0.00");
-                SetTerrainEnabled();
-            }
-        }
-
-        private void Types_eventCheckChanged(UIComponent component, bool value)
-        {
-            if (plRoads.isVisible)
-            {
-                SetUpdateEnabled();
-            }
-            else
-            {
-                SetDeleteEnabled();
-            }
-        }
-
-        private void ServiceTypes_eventCheckChanged(UIComponent component, bool value)
-        {
-
-            //do we reenble or disable chirper
-            UICheckBox cb = (UICheckBox)component;
-            
-            //WriteLog("Entering ServiceTypes_eventCheckChanged.Name, Value: " + cb.text + ", Value: " + value);
-
-            if (cb.text == "Chirper")
-            {
-                //we need to toggle shown or not
-                if (Chirp.ToggleState != value)
-                    Chirp.Toggle(value);
-                return;
-            }
-
-            if (cb.text == "Abandoned")
-            {
-                //Toggle tha value as needed
-                if (DemolishAbandoned.@value != value)
-                    DemolishAbandoned.@value = !DemolishAbandoned.@value;
-                return;
-            }
-
-            if (cb.text == "Burned")
-            {
-                //Toggle tha value as needed
-                if (DemolishBurned.@value != value)
-                    DemolishBurned.@value = !DemolishBurned.@value;
-                return;
-            }
-
-            SetServicesEnabled();
-            //WriteLog("Leaving ServiceTypes_eventCheckChanged");
-        }
-
-        private void btHelp_eventDoubleClick(UIComponent component, UIMouseEventParameter eventParam)
-        {
-            //WriteLog("Entering btHelp_eventDoubleClick");
-
-            try
-            {
-                Assembly assembly = Assembly.GetExecutingAssembly();
-                Stream stream = assembly.GetManifestResourceStream("AnotherRoadUpdate.ARUT.pdf");
-                if (stream == null)
-                {
-                    //WriteLog("Error loading embeded resource AnotherRoadUpdate.ARUT.pdf");
-                    return;
-                }
-                BinaryReader br = new BinaryReader(stream);
-                FileStream fs = new FileStream("ARUT.pdf", FileMode.Create);
-                BinaryWriter bw = new BinaryWriter(fs);
-                byte[] ba = new byte[stream.Length];
-                stream.Read(ba, 0, ba.Length);
-                bw.Write(ba);
-                br.Close();
-                bw.Close();
-                Process p = new Process();
-                p.StartInfo.FileName = "ARUT.pdf";
-                p.Start();
-
-            }
-            catch (Exception ex)
-            {
-                WriteError("Error in btHelp_eventDoubleClick ", ex);
-            }
-
-            //WriteLog("Leaving btHelp_eventDoubleClick");
-        }
-
-        private void DeleteTypes_eventCheckChanged(UIComponent component, bool value)
-        {
-            if (m_updates == true) { return; }
-            m_updates = true;
-
-            //WriteLog("Entering DeleteTypes_eventCheckChanged");
-            UICheckBox c = (UICheckBox)component;
-            //store the update
-            if (c.text == deletes[(int)p.Roads].text) { us.Roads = value; }
-            if (c.text == deletes[(int)p.Railroads].text) { us.Railroads = value; }
-            if (c.text == deletes[(int)p.Highways].text) { us.Highways = value; }
-            if (c.text == deletes[(int)p.PowerLines].text) { us.PowerLines = value; }
-            if (c.text == deletes[(int)p.WaterPipes].text) { us.WaterPipes = value; }
-            if (c.text == deletes[(int)p.HeatPipes].text) { us.HeatPipes = value; }
-            if (c.text == deletes[(int)p.Airplanes].text) { us.Airplanes = value; }
-            if (c.text == deletes[(int)p.Shipping].text) { us.Shipping = value; }
-
-            if (c.text == deletes[(int)p.Buildings].text) { us.Buildings = value; }
-            if (c.text == deletes[(int)p.Props].text) { us.Props = value; }
-            if (c.text == deletes[(int)p.Trees].text) { us.Trees = value; }
-
-            if (c.text == types[(int)tp.Ground].text) { us.Ground = value; }
-            if (c.text == types[(int)tp.Bridge].text) { us.Bridge = value; }
-            if (c.text == types[(int)tp.Tunnel].text) { us.Tunnel = value; }
-            if (c.text == types[(int)tp.Slope].text) { us.Tunnel = value; }
-            if (c.text == types[(int)tp.Curve].text) { us.Tunnel = value; }
-
-            //set enabled
-            SetDeleteEnabled();
-            m_updates = false;
-            //WriteLog("Leaving DeleteTypes_eventCheckChange Roads: " + us.Roads + " Ground: " + us.Ground);
-        }
-
-        private void Options_eventCheckChanged(UIComponent component, bool value)
-        {
-            if (m_updates == true) { return; }
-            m_updates = true;
-
-            UICheckBox c = (UICheckBox)component;
-            //WriteLog("Entering: Options_eventCheckChanged: c.text: " + c.text + " Checked: " + c.isChecked);
-
-            //Hide tham all
-            UpdatePanels(false);
-
-            //did we unclick tham all?
-            if (c.isChecked == false) { }
-            else
-            {
-                if (c.text == "Updates")
-                {
-                    options[(int)ops.Deletes].isChecked = false;
-                    options[(int)ops.Services].isChecked = false;
-                    options[(int)ops.Terrain].isChecked = false;
-                    options[(int)ops.Districts].isChecked = false;
-                    plTypes.isVisible = true;
-                    plRoads.isVisible = true;
-                    ShowRoads();
-                }
-                else if (c.text == "Deletes")
-                {
-                    options[(int)ops.Updates].isChecked = false;
-                    options[(int)ops.Services].isChecked = false;
-                    options[(int)ops.Terrain].isChecked = false;
-                    options[(int)ops.Districts].isChecked = false;
-                    plTypes.isVisible = true;
-                    plDelete.isVisible = true;
-                    SetDeleteEnabled();
-                }
-                else if (c.text == "Services")
-                {
-                    options[(int)ops.Updates].isChecked = false;
-                    options[(int)ops.Deletes].isChecked = false;
-                    options[(int)ops.Terrain].isChecked = false;
-                    options[(int)ops.Districts].isChecked = false;
-                    plTypes.isVisible = false;
-                    plServices.isVisible = true;
-                    SetServicesEnabled();
-                }
-                else if (c.text == "Terrain")
-                {
-                    options[(int)ops.Updates].isChecked = false;
-                    options[(int)ops.Deletes].isChecked = false;
-                    options[(int)ops.Services].isChecked = false;
-                    options[(int)ops.Districts].isChecked = false;
-                    plTypes.isVisible = false;
-                    plTerrain.isVisible = true;
-                    SetTerrainEnabled();
-                }
-                else if (c.text == "Districts")
-                {
-                    options[(int)ops.Updates].isChecked = false;
-                    options[(int)ops.Deletes].isChecked = false;
-                    options[(int)ops.Services].isChecked = false;
-                    options[(int)ops.Terrain].isChecked = false;
-                    plTypes.isVisible = false;
-                    plDistricts.isVisible = true;
-                    SetDistrictsEnabled();
-                }
-            }
-            m_updates = false;
-            //WriteLog("Leaving Options_eventCheckChanged");
-        }
-
-        private void FromTypes_eventCheckChanged(UIComponent component, bool value)
-        {
-            if (m_fromTypes == true) { return; }
-            m_fromTypes = true;
-
-            UICheckBox cb = (UICheckBox)component;
-
-            //WriteLog("Entering FromTypes_eventCheckChanged: Option: " + cb.text);
-
-            int loc = 0;
-            try
-            {
-                plBasic.isVisible = false;
-                plHighway.isVisible = false;
-                plLarge.isVisible = false;
-                plMedium.isVisible = false;
-                plOneway.isVisible = false;
-                loc += 1;
-            }
-            catch (Exception ex)
-            {
-                WriteError("Error in FromTypes_eventCheckChanged loc: " + loc + ".", ex);
-            }
-
-            //uncheck any others
-            foreach (UICheckBox c in fromTypes)
-            {
-                if (c.text != cb.text) { c.isChecked = false; }
-            }
-
-            if (cb.isChecked == true)
-            {
-                if (cb.text == "Basic") { plBasic.isVisible = true; }
-                if (cb.text == "Highway") { plHighway.isVisible = true; }
-                if (cb.text == "Large") { plLarge.isVisible = true; }
-                if (cb.text == "Medium") { plMedium.isVisible = true; }
-                if (cb.text == "Oneway") { plOneway.isVisible = true; }
-            }
-
-            m_fromTypes = false;
-
-            //WriteLog("Leaving FromTypes_eventCheckChanged: Option: " + cb.text);
-        }
-
-        private void ToTypes_eventCheckChanged(UIComponent component, bool value)
-        {
-            if (m_toTypes == true) { return; }
-            m_toTypes = true;
-
-            UICheckBox cb = (UICheckBox)component;
-
-            //WriteLog("Entering ToTypes_eventCheckChanged: Option: " + cb.text);
-
-            int loc = 0;
-            try
-            {
-                plToBasic.isVisible = false;
-                plToHighway.isVisible = false;
-                plToLarge.isVisible = false;
-                plToMedium.isVisible = false;
-                plToOneway.isVisible = false;
-                loc += 1;
-            }
-            catch (Exception ex)
-            {
-                WriteError("Error in ToTypes_eventCheckChanged loc: " + loc + ".", ex);
-            }
-
-            //uncheck any others
-            foreach (UICheckBox c in toTypes)
-            {
-                if (c.text != cb.text) { c.isChecked = false; }
-            }
-
-            if (cb.isChecked == true)
-            {
-                if (cb.text == "Basic") { plToBasic.isVisible = true; }
-                if (cb.text == "Highway") { plToHighway.isVisible = true; }
-                if (cb.text == "Large") { plToLarge.isVisible = true; }
-                if (cb.text == "Medium") { plToMedium.isVisible = true; }
-                if (cb.text == "Oneway") { plToOneway.isVisible = true; }
-            }
-
-            m_toTypes = false;
-
-            //WriteLog("Leaving ToTypes_eventCheckChanged: Option: " + cb.text);
-        }
-
-        private void FromRoad_eventCheckChanged(UIComponent component, bool value)
-        {
-            if (m_fromRoads == true) { return; }
-            m_fromRoads = true;
-
-            UICheckBox cb = (UICheckBox)component;
-
-            //WriteLog("Entering FromRoads_eventCheckChanged: Option: " + cb.text);
-
-            //uncheck any others
-            foreach (UICheckBox c in fromOneway) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in fromMedium) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in fromHighway) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in fromLarge) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in fromBasic) { if (c.text != cb.text) { c.isChecked = false; } }
-
-            fromSelected = String.Empty;
-            if (cb.isChecked) { fromSelected = cb.text; }
-
-            SetUpdateEnabled();
-
-            m_fromRoads = false;
-
-            //WriteLog("Leaving FromRoads_eventCheckChanged: Option: " + cb.text);
-        }
-
-        private void ToRoad_eventCheckChanged(UIComponent component, bool value)
-        {
-            //WriteLog("Entering ToRoads_eventCheckChanged: m_toRoads: " + m_toRoads);
-            if (m_toRoads == true) { return; }
-            m_toRoads = true;
-
-            UICheckBox cb = (UICheckBox)component;
-
-            //WriteLog("Entering ToRoads_eventCheckChanged: Option: " + cb.text);
-
-            //uncheck any others
-            foreach (UICheckBox c in toOneway) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in toMedium) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in toHighway) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in toLarge) { if (c.text != cb.text) { c.isChecked = false; } }
-            foreach (UICheckBox c in toBasic) { if (c.text != cb.text) { c.isChecked = false; } }
-
-            toSelected = String.Empty;
-            if (cb.isChecked) { toSelected = cb.text; }
-
-            SetUpdateEnabled();
-
-            m_toRoads = false;
-
-            //WriteLog("Leaving ToRoads_eventCheckChanged: Option: " + cb.text);
-        }
-
+        
         private void Button_Clicked(UIComponent component, UIMouseEventParameter eventParam)
         {
-            //WriteLog("Entering Button_Clicked: " + plMain.isVisible);
-            if (plMain.isVisible == true)
-            {
-                this.OnDisable();
-                plMain.isVisible = false;
-                this.enabled = false;
-            }
-            else
-            {
-                this.enabled = true;
-                plMain.isVisible = true;
-                mainButton.enabled = true;
-            }
-            //WriteLog("Leaving Button_Clicked: " + plMain.isVisible);
+            plMain.isVisible = !plMain.isVisible;
         }
 
         #endregion
@@ -2372,24 +721,6 @@ namespace AnotherRoadUpdateTool
             return true;
         }
 
-        private bool ValidateSelectedArea(Building bd)
-        {
-            var minX = this.m_startPosition.x < this.m_mousePosition.x ? this.m_startPosition.x : this.m_mousePosition.x;
-            var minZ = this.m_startPosition.z < this.m_mousePosition.z ? this.m_startPosition.z : this.m_mousePosition.z;
-            var maxX = this.m_startPosition.x > this.m_mousePosition.x ? this.m_startPosition.x : this.m_mousePosition.x;
-            var maxZ = this.m_startPosition.z > this.m_mousePosition.z ? this.m_startPosition.z : this.m_mousePosition.z;
-
-            Vector3 position = bd.m_position;
-
-            float positionDiff = Mathf.Max(Mathf.Max(minX - 16f - position.x, minZ - 16f - position.z), Mathf.Max(position.x - maxX - 16f, position.z - maxZ - 16f));
-
-            if (positionDiff < 0f)
-            {
-                return true;
-            }
-            return false;
-        }
-        
         private static float ConvertCoords(float coords, bool ScreenToTerrain = true)
         {
             return ScreenToTerrain ? coords / 16f + 1080 / 2 : (coords - 1080 / 2) * 16f;
@@ -2468,6 +799,7 @@ namespace AnotherRoadUpdateTool
             {
                 byte id;
                 bool result = dm.CreateDistrict(out id);
+
                 dm.AreaModified(minX, minZ, maxX, maxZ, true);
             }
             catch (Exception ex)
@@ -2475,71 +807,11 @@ namespace AnotherRoadUpdateTool
                 WriteError("Error in ApplyDistrictsChange: ", ex);
             }
         }
-
-        private void ApplyServices()
-        {
-            //WriteLog("Entering ApplyServices: Code not yet implamented.");
-            Building[] buffer = Singleton<BuildingManager>.instance.m_buildings.m_buffer;
-
-            bool toggle = cbToggle.isChecked;
-            //string onoff = "Off";
-            byte rate = 0;
-
-
-            for (int i = 0; i < buffer.Length; i++)
-            {
-                Building bd = buffer[i];
-
-                if (services[(int)dl.Beautification].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.Beautification) { }
-                else if (services[(int)dl.Education].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.Education) { }
-                else if (services[(int)dl.Electricity].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.Electricity) { }
-                else if (services[(int)dl.FireDepartment].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.FireDepartment) { }
-                else if (services[(int)dl.Garbage].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.Garbage) { }
-                else if (services[(int)dl.HealthCare].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.HealthCare) { }
-                else if (services[(int)dl.Monument].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.Monument) { }
-                else if (services[(int)dl.PoliceDepartment].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.PoliceDepartment) { }
-                else if (services[(int)dl.PublicTransport].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.PublicTransport) { }
-                else if (services[(int)dl.Water].isChecked == false && bd.Info.m_class.m_service == ItemClass.Service.Water) { }
-                else if (ValidateSelectedArea(bd) == false)
-                {
-                    //sw.WriteLine(String.Format("Segment {0} is not in the selected area.", segment.Info.name));
-                }
-                else
-                {
-                    //WriteLog("Building Name: " + bd.Info.gameObject.name + " GetName: " + GetName(bd));
-                    //WriteLog("BuildingAI Name: " + bd.Info.m_buildingAI.name);
-                    //WriteLog("Building.Info.GetLocalizedTitle : " + bd.Info.GetLocalizedTitle());
-                    
-                    //onoff = "Off";
-                    rate = 0;
-                    if (toggle == true)
-                    {
-                        //onoff = "On";
-                        rate = 100;
-                    }
-
-                    //WriteLog("Setting services to " + onoff + ". bd.m_flags was: " + bd.m_flags.IsFlagSet(Building.Flags.Active));
-
-                    if (bd.m_flags.IsFlagSet(Building.Flags.Active) != toggle)
-                    {
-                        //WriteLog("About to toggle: ");
-                        bd.m_flags ^= Building.Flags.Active;
-                        bd.m_productionRate = rate;
-                    }
-
-                    //WriteLog("Setting services to " + onoff + ". bd.m_flags is: " + bd.m_flags.IsFlagSet(Building.Flags.Active));
-                }
-                buffer[i] = bd;
-            }
-            //WriteLog("Leaving ApplyServices: Code not yet implamented.");
-        }
-
+        
         private void ApplyUndo()
         {
-            if (UndoList.Count < 0)
-            {
-                return;
-            }
+            if (UndoList.Count <= 0) { return; }
+            WriteLog("Entring ApplyUndo");
             //remove the current changes from the list (there are none)
             UndoStroke undoStroke = UndoList[UndoList.Count - 1];
             UndoList.RemoveAt(UndoList.Count - 1);
@@ -2586,7 +858,8 @@ namespace AnotherRoadUpdateTool
                 i += 119;
                 if (i > maxZ)
                     i = maxZ - 1;
-            }            
+            }
+            WriteLog("Leaving ApplyUndo");
         }
 
         #endregion
